Refactor: extract UploaderPathSecurity shared module
Move the upload-path canonicalization guard, same_site_url? and
site_url_path out of both RuntimeUploaderConcern and UploaderHelper into a
single CamaleonCms::UploaderPathSecurity module included by both, so the
four upload-sink guards and same-site URL detection can no longer drift
between the two copies. Pure refactor — behavior unchanged.
