+++ app/controllers/concerns/camaleon_cms/runtime_uploader_concern.rb
@@ -3,10 +3,11 @@ require 'tempfile'
 
 # rubocop:disable Naming/MethodParameterName
 module CamaleonCms
-  module RuntimeUploaderConcern # rubocop:disable Metrics/ModuleLength
+  module RuntimeUploaderConcern
     extend ActiveSupport::Concern
 
     include UploaderContentSecurity
+    include UploaderPathSecurity
 
     def upload_file(uploaded_io, settings = {})
       cached_name = uploaded_io.is_a?(ActionDispatch::Http::UploadedFile) ? uploaded_io.original_filename : nil
@@ -20,15 +21,8 @@ module CamaleonCms
         uploaded_io = tmp[:file_path]
       end
       if uploaded_io.is_a?(String)
-        expanded = begin
-          File.expand_path(uploaded_io)
-        rescue ArgumentError, TypeError
-          return { error: 'Invalid file path' }
-        end
-        roots = [Rails.public_path.to_s, Dir.tmpdir]
-        unless roots.any? { |r| expanded == r || expanded.start_with?(r + File::SEPARATOR) }
-          return { error: 'Invalid file path' }
-        end
+        expanded = cama_canonical_upload_path(uploaded_io)
+        return { error: 'Invalid file path' } unless expanded
 
         uploaded_io = expanded
       end
@@ -197,7 +191,7 @@ module CamaleonCms
       file
     end
 
-    def cama_tmp_upload(uploaded_io, args = {}) # rubocop:disable Metrics/PerceivedComplexity
+    def cama_tmp_upload(uploaded_io, args = {})
       tmp_path = args[:path] || File.join(Rails.public_path, 'tmp', current_site.id.to_s).to_s
       FileUtils.mkdir_p(tmp_path)
       saved = false
@@ -234,15 +228,8 @@ module CamaleonCms
         args[:name] = args[:name] || _tmp_name
       end
       if uploaded_io.is_a?(String)
-        expanded = begin
-          File.expand_path(uploaded_io)
-        rescue ArgumentError, TypeError
-          return { error: 'Invalid file path' }
-        end
-        roots = [Rails.public_path.to_s, Dir.tmpdir]
-        unless roots.any? { |r| expanded == r || expanded.start_with?(r + File::SEPARATOR) }
-          return { error: 'Invalid file path' }
-        end
+        expanded = cama_canonical_upload_path(uploaded_io)
+        return { error: 'Invalid file path' } unless expanded
 
         uploaded_io = expanded
       end
@@ -389,24 +376,6 @@ module CamaleonCms
       return if cama_uploader.class.validate_file_format(file, formats)
 
       { error: "#{I18n.t('camaleon_cms.common.file_format_error')} (#{formats})" }
-    end
-
-    def same_site_url?(url, site)
-      uri = URI.parse(url)
-      site_uri = URI.parse(site.the_url(locale: nil))
-      uri.host == site_uri.host && uri.port == site_uri.port
-    rescue URI::InvalidURIError
-      false
-    end
-
-    def site_url_path(url, site)
-      uri = URI.parse(url)
-      path = uri.path
-      langs = site.get_languages
-      path = path.sub(%r{\A/(?:#{Regexp.union(langs.map(&:to_s))})(?=/|$)}, '') if langs.size > 1
-      path
-    rescue URI::InvalidURIError
-      url
     end
   end
 end
